+++ src/app.ts
@@ -2,7 +2,7 @@ import dotenv from 'dotenv';
 import express from 'express';
 import connectDB from './config/db';
 import AuthRouter from './routes/AuthRoutes';
-import SocialRoutes from './routes/SocialAuthRoutes';
+import SocialRoutes from './routes/SocialLoginRoutes';
 import ProductRouter from './routes/ProductRoutes';
 import cors from 'cors';
