+++ src/routes/AuthRoutes.ts
@@ -1,11 +1,10 @@
 import express from 'express';
-import { emailLogin, emailSignUp, socialLogin, verifyEmail } from '../controllers/AuthController';
+import { emailLogin, emailSignUp, verifyEmail } from '../controllers/AuthController';
 
 const router = express.Router();
 
 router.post('/signup', emailSignUp);
 router.get('/verify-email', verifyEmail);
 router.post('/login', emailLogin);
-router.post('/social-login', socialLogin);
 
 export default router;
