+++ src/services/NaverLoginService.ts
@@ -1,0 +1,38 @@
+import axios from 'axios';
+import { findOrCreateUser } from '../repositories/NaverLoginRepository';
+import dotenv from 'dotenv';
+dotenv.config();
+
+export class SocialAuthService {
+   // 네이버 로그인 URL 생성
+   static getNaverAuthUrl(): string {
+      const clientId = process.env.NAVER_CLIENT_ID || '';
+      const redirectUri = encodeURIComponent(`${process.env.BACKEND_URL}/socialLogin/naver/callback`);
+      const state = 'RANDOM_STATE';
+      return `https://nid.naver.com/oauth2.0/authorize?response_type=code&client_id=${clientId}&redirect_uri=${redirectUri}&state=${state}`;
+   }
+
+   // 네이버 콜백 처리
+   static async handleNaverCallback(code: string, state: string) {
+      const clientId = process.env.NAVER_CLIENT_ID || '';
+      const clientSecret = process.env.NAVER_CLIENT_SECRET || '';
+      const redirectUri = `${process.env.BACKEND_URL}/socialLogin/naver/callback`;
+
+      try {
+         const tokenUrl = `https://nid.naver.com/oauth2.0/token?grant_type=authorization_code&client_id=${clientId}&client_secret=${clientSecret}&redirect_uri=${redirectUri}&code=${code}&state=${state}`;
+         const tokenResponse = await axios.get(tokenUrl);
+         const { access_token } = tokenResponse.data;
+
+         const userInfoUrl = 'https://openapi.naver.com/v1/nid/me';
+         const userInfoResponse = await axios.get(userInfoUrl, {
+            headers: { Authorization: `Bearer ${access_token}` },
+         });
+
+         // 유저 정보 저장 및 반환
+         return findOrCreateUser('naver', userInfoResponse.data);
+      } catch (error) {
+         console.error('네이버 로그인 처리 중 오류:', error);
+         throw new Error('네이버 로그인 처리 중 오류 발생');
+      }
+   }
+}
